Moved UninterruptibleHashTable.
